requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
torchmetrics>=1.3.0
pycocotools>=2.0.7
numpy>=1.25.0
pandas>=2.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pillow>=10.0.0
tensorboard>=2.14.0
tqdm>=4.66.0
scipy>=1.11.0
seaborn>=0.12.2
requests
flask